+++ src/main/java/com/careerdevs/gorestfinal/repositories/CommentRepository.java
@@ -3,5 +3,5 @@ package com.careerdevs.gorestfinal.repositories;
 import com.careerdevs.gorestfinal.models.Comment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommentRepository extends CrudRepository<Comment, Integer> {
+public interface CommentRepository extends CrudRepository<Comment, Long> {
 }
