+++ src/main/java/com/careerdevs/gorestfinal/repositories/ToDoRepository.java
@@ -3,5 +3,5 @@ package com.careerdevs.gorestfinal.repositories;
 import com.careerdevs.gorestfinal.models.ToDo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ToDoRepository extends CrudRepository<ToDo, Integer> {
+public interface ToDoRepository extends CrudRepository<ToDo, Long> {
 }
